chore(Menu) : 메뉴 관련 enum 클래스를 model 패키지로 이동
